requirements.txt
aiogram
python-dotenv
yt-dlp
imageio-ffmpeg